requirements_old.txt
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.2
docker==2.3.0
wharfee==0.10
